Refactor email obfuscation logic in index.html
The JavaScript code in index.html has been modified to improve the obfuscation of the email address on the webpage. The function `obfuscateEmail` now takes two parameters, `user` and `domain`, and generates an obfuscated email address by converting each character into its corresponding HTML entity representation. The obfuscated email is then displayed on the webpage using the `innerHTML` property of the element with the id `obfuscatedEmail`. This change enhances the security and prevents email harvesting.